Support browser() as a thenable rejection
