+++ packages/react-reconciler/src/ReactChildFiber.js
@@ -38,7 +38,6 @@ import {
   REACT_PORTAL_TYPE,
   REACT_LAZY_TYPE,
   REACT_CONTEXT_TYPE,
-  REACT_RECOVERABLE_TYPE,
   REACT_LEGACY_ELEMENT_TYPE,
   REACT_OPTIMISTIC_KEY,
 } from 'shared/ReactSymbols';
@@ -825,13 +824,6 @@ function createChildReconciler(
         );
       }
 
-      // $FlowFixMe[invalid-compare]
-      if (newChild.$$typeof === REACT_RECOVERABLE_TYPE) {
-        // Fiber is the final renderer, so there is no downstream host that
-        // needs to recover this subtree. Treat it as an empty child.
-        return null;
-      }
-
       throwOnInvalidObjectType(returnFiber, newChild);
     }
 
@@ -975,13 +967,6 @@ function createChildReconciler(
         );
       }
 
-      // $FlowFixMe[invalid-compare]
-      if (newChild.$$typeof === REACT_RECOVERABLE_TYPE) {
-        // Fiber is the final renderer, so there is no downstream host that
-        // needs to recover this subtree. Treat it as an empty child.
-        return null;
-      }
-
       throwOnInvalidObjectType(returnFiber, newChild);
     }
 
@@ -1114,13 +1099,6 @@ function createChildReconciler(
           readContextDuringReconciliation(returnFiber, context, lanes),
           lanes,
         );
-      }
-
-      // $FlowFixMe[invalid-compare]
-      if (newChild.$$typeof === REACT_RECOVERABLE_TYPE) {
-        // Fiber is the final renderer, so there is no downstream host that
-        // needs to recover this subtree. Treat it as an empty child.
-        return null;
       }
 
       throwOnInvalidObjectType(returnFiber, newChild);
@@ -2019,13 +1997,6 @@ function createChildReconciler(
           readContextDuringReconciliation(returnFiber, context, lanes),
           lanes,
         );
-      }
-
-      // $FlowFixMe[invalid-compare]
-      if (newChild.$$typeof === REACT_RECOVERABLE_TYPE) {
-        // Fiber is the final renderer, so there is no downstream host that
-        // needs to recover this subtree. Treat it as an empty child.
-        return deleteRemainingChildren(returnFiber, currentFirstChild);
       }
 
       throwOnInvalidObjectType(returnFiber, newChild);
