+++ packages/react-server/src/ReactFizzServer.js
@@ -15,7 +15,6 @@ import type {
 import type {
   ReactNodeList,
   ReactContext,
-  ReactRecoverable,
   ReactConsumerType,
   Wakeable,
   Thenable,
@@ -135,10 +134,12 @@ import {
   readPreviousThenableFromState,
   getActionStateCount,
   getActionStateMatchingIndex,
+} from './ReactFizzHooks';
+import {
   createRecoverableError,
   isRecoverableError,
   cloneRecoverableErrorAsFatal,
-} from './ReactFizzHooks';
+} from './ReactFizzRecoverable';
 import {DefaultAsyncDispatcher} from './ReactFizzAsyncDispatcher';
 import {
   getStackByComponentStackNode,
@@ -3719,12 +3720,6 @@ function retryNode(request: Request, task: Task): void {
         readContext(context),
         childIndex,
       );
-    }
-
-    if (maybeUsable.$$typeof === REACT_RECOVERABLE_TYPE) {
-      // Materialize the diagnostic when the rendered usable is consumed.
-      const recoverable: ReactRecoverable = maybeUsable as any;
-      throw createRecoverableError(recoverable);
     }
 
     // $FlowFixMe[method-unbinding]
